Rename the image-flip css selector.
Instead of a blanket `img.hieroglyphic` class selector to trigger
the image flip so the left-to-right rendered pngs align with the
right-to-left order of the heiratic, use the more semantically-
accurate `flip` class name, so it's more clear what's going on.

Also run `python generate_webpage.py` to update the class names
on the generated html.