Use shutil.copyfile to place images into docs.
Instead of opening the image and writing it out again, copy
the original image into the website tree. This makes the images
identical, so it's easier to track changes. It should also
save space in the git repository, since it doesn't have to
track two compressed versions.

Also, define the filename pattern in a variable so it can be
changed in a single place.