Transcribe hrw as written on line 5.10.
Gardiner has  𓉔𓂋𓏲𓇳𓏤which seems to be a correction to the Middle
Egyptian form of the word. Spellings similar to 𓉔𓏹𓏹𓇳𓏤 are attested
in Lesko and seem like a much better match for what's actually
written here.

Aurelio suggests that the /r/ has probably become of glottal stop
by this point, *haɂw > Coptic ϩⲟⲟⲩ, so the 𓂋 is reduced to a stroke.