Transcribe 𓎛𓀜𓏷𓀜 with Z12 instead of G17
Based on today's discussion, follow Gardiner, Lesko, Ramses
Online, and Enzo Chiurco by transcribing this sign as 𓏷 even
though it looks a lot like 𓅓 in our papyrus. Other scribes
use a form much closer to the jsesh/unicode shape,
e.g. Anastasi III 3.13.

Z12 𓏷 seems to be a hieratic sign for which we don't know a
hieroglyphic equivalent. Gardiner and Faulkner list the Middle
Egyptian version of the same verb without the sign at all in
their dictionaries.

Looking closely, especially at the instance on line 8.13,
there may be some difference with the '3' glyph used by
the scribe for 𓅓 in that the upper curve is more angular,
the sign is wider, and the lower curve starts on the baseline.