Fix transcription after cartouche in line 1.6
We had 𓏙𓋹𓍑𓋴 after Wennefer's cartouche, but that doesn't make
sense. 𓏙𓋹 is a phrase and 𓋹𓍑𓋴 is a phrase, but running them
together isn't. But there's definitely another vertical stroke
between the finish of the cartouche and the l.p.h. abbreviation.

Maybe it's a mistake, or short for some other symbol. Follow
Gardiner and Ramses online by transcribing it as FF1, although
it looks as much like as Z1 as anything else.

Also flip the reading direction to align with the hieratic text
like the scripts expect for newly-transcribed lines.